Use two interfaces: one for Nova and one for Glance
Also, mocking approach is now updated so that mock base
implements all functions of the interface, but in a way
that it returns error wherever possible. Each unit test
then either extends mock base and implements only functions
that it really needs, or uses the errorous mock directly to
check unhappy path.

Change-Id: I565bb5ffa6380bfc3fb726fc5016a39db39a69f7
Signed-off-by: Miha Pleško <miha.plesko@xlab.si>