Let MPM package remember creation time
Now that capstan-packages building platform is available we
come to situation when we start wondering whether we have the
latest MPM packages in our local repository.

With this commit we replace field ModTime with Create where we
store timestamp of the moment when `capstan package init` was
called. This way user is able to differentiate between multiple
builds of the same version of the package. Mostly useful for
debugging.

Change-Id: I42befbffcf45f77a4ce77ca593c8c0e0f199e5ea
Signed-off-by: Miha Pleško <miha.plesko@xlab.si>