Introduce 'platform' attribute in meta/package.yaml
With this commit we modify three capstan commands:

- capstan package init   -> we introduce --platform argument
- capstan package list   -> we add a column to the printed result
- capstan package search -> we add a column to the printed result

The 'platform' attribute is a string e.g. "Ubuntu-14.04" that tells us
what platform was the package built on.

In order to be able to unut test the modified output of
'capstan package list' we quickly refactor the ListPackages function
so that it now returns string (rather than print directly to console).

Change-Id: Ie47d9bdd3f05d1accf676aa1a5aafc5c63f91c01
Signed-off-by: Miha Pleško <miha.plesko@xlab.si>