Add 'platform' column to the output of `capstan images`
It's useful to know what platform was image built on (e.g. "Ubuntu-14.04"),
especially in case of mike/osv-loader since only packages built on compatible
platform will work there. With this commit we tried to just add additional
attribute to the index.yaml of each image. BUT it turned out that some
refactoring is needed in order for the code to become testable.

Namely, we've renamed "MakeFileInfo" function into "ParseIndexYaml" since the
former name does not make any sense. Also we reimplement its caller,
the "ListImages" function, so that it now returns string instead writing
directly to console, and write unit tests for it.

Change-Id: Ie2697a16b054cd827d03395447a663c3a853e5fc
Signed-off-by: Miha Pleško <miha.plesko@xlab.si>